face_matching/requirements.txt
torch
facenet-pytorch
numpy
opencv-python
scipy
scikit-learn
Pillow
tqdm
python-speech-features
pandas
